Add files via upload
it is a sample ATM SIMULATION. It starts by importing the time module and initializing the PIN, balance, and transaction history. The user is prompted to insert their card, and the program waits for 5 seconds. Inside an infinite loop, the user is given choices to check balance, withdraw cash, deposit cash, change the PIN, view transaction history, or exit. Each operation requires PIN verification. The balance and transaction history are updated accordingly, with error handling for invalid inputs. The loop continues until the user chooses to exit.